renames helling module
